requirements.txt
torch
transformers>=5.1.0
datasets
numpy
pandas
scikit-learn
safetensors
tqdm
scipy
accelerate
pyyaml
tiktoken
protobuf
sentencepiece
